I'm sorry God, here is an offering
No one's buying your product because of your social messaging, they're buying it because its good. Now apply this to open source software.

"Have you ever thought that maybe certain things inside of your company or inside of your system maybe they don't need to exist to begin with"

"An idiot admires complexity, A genius simplicity"

and on that note back to the hole