under a star called sun
the trajectory of this ship is

unchanging

i check on the plants

in the greenhouse

i check for bugs

i check the moisture of the soil

i check the dust on the leaves

And when no one’s around

i hum to them

in a normal message

i’d be saying shit like

how’s it going

what’s new

so when are you finally gonna move to melbourne

im so excited to see you again

oh atlas

holds up the firmament

and endures

one day at a time

i remember when i first

heard the news

i woke up,

earlier than usual

and i get a message from our friend

and my phone freezes

and one could attribute this to

planned obsolescence

but maybe, maybe, maybe...

you know, there’s a room

on this ship

(i wont tell you how it works)

i can replay the moment

i last saw you

over and over and over and over

it was late morning on a

perfectly ordinary august day

we were meeting up for brunch

at a cafe close to mine

a line up of friends

i only see occasionally

i think you ordered

a fancy sandwich

or maybe just a salad

then we sat in the park and

talked for a while

you had to get going

meeting another friend

you stepped on the 19 tram

to north coburg and smiled

framed between

the pneumatic doors

beautiful under

the fluorescent light

you said see you next time

was it sunny or cloudy that day?

what did you order?

what did we talk about?

what did your voice sound like?

what was the expression on your

face when we said goodbye?